Remove unused-variable in fboss/agent/test/LookupClassUpdaterTests.cpp +3
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: shri-khare

Differential Revision: D66978890

fbshipit-source-id: 2c42c6a80f1e085bd9166f653bb72afae8f4df76