Remove unused-variable in fboss/lib/test/PciDeviceTest.cpp +3
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: harshitgulati18

Differential Revision: D66978877

fbshipit-source-id: 3c681cc81fb78727ad2a0941acfc7c88c9a347a4