Remove unused-variable in fboss/agent/hw/sai/hw_test/HwTestFlowletSwitchingUtils.cpp +3
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: msomasundaran

Differential Revision: D66978908

fbshipit-source-id: 73b8d4d462e6db19a7cc586d52affd8803472dea