Improve SpiDevice error handling
Summary:
We have a confusing error because when SpiDevice fails, we treat that as its SpiMaster failed.
So if SpiDevice X fails, in error reporting it will be reported as that SpiMaster failed.

We create a symlink for each SpiDevice and it will fail if SpiDevice creation fails.
So without some context, kinda hard to grasp a reason for symlink failure because the SpiDevice creation failure would be reported under its SpiMaster device path.

Reviewed By: somasun

Differential Revision: D67031958

fbshipit-source-id: 0737a252786d3f1bfe4987c8075e2a78dd5f666a