Add test for asserting link active, inactive status in drain
Summary:
Previously semantics were that active status depended
only on peer port/device drain status. However this was
recently changed to consider local drain status as well.
So if port is drained locally, it also turns inactive, regardless
of peer drain status.
Assert that in a test. Note this is a link test since, in HW test
peer port and local port are the same so we can't differentiate;

Reviewed By: nivinl

Differential Revision: D67156889

fbshipit-source-id: 16ee8f3293212d4f0027180d9c05c30f86028b93