Remove sd_notify and let watchdog handle it
Summary:
When a MetalOS Native Service is provisioned several things are happening to bring up the environment.

Here are the several drop in service units created along aside the primary one:
Example: platform_manager
**1. metalos-native-service-certs@metalos.wds.platform_manager.service**
 Responsible for fetching tlscerts from tlsrcertreq/Root Canal service prior to the primary service starting

**2. metalos-native-service-watchdog@metalos.wds.platform_manager.service**
 Responsible for the running a custom watchdog service binary that the standard implementation is using the fb303 health checks. This is responsible for sending the sd_notify API once the default fb303 health checks have passed. This moves the entire service into an ACTIVE state in systemd for subsequent services to come up.

**3. metalos-native-service@metalos.wds.platform_manager.service**
Responsible for bringing up the custom filesystem that contains the primary service unit's binary and metadata/artifacts to launch the service. This will come up and hold the BTRFS volume up and go to sleep forever after provisioned.

Due to the fact that the Platform_Manager and the Service Unit 2 from above are both sending the sd_notify signal it won't be clear if the service is truly healthy. As a result, to have a single source of truth the NetOS Platform Manager will disable sending sd_notify in this environment.

P1666383160

Ref: https://www.internalfb.com/intern/staticdocs/metalos/docs/os-components/watchdogs/#watchdog-unit

Reviewed By: matthazinski

Differential Revision: D64925022

fbshipit-source-id: 3c1da6903a5e95407a381778e97c51ba2a729605