DLB: SAI switch impl for NHG
Summary:
This is the core change to enable DLB on NHG. There are 2 paths where NHG is updated, creation and during config updates.
1. During creation, add the ARS object attribute
2. During config update, all NHGs are checked and converted to DLB by setting the ARS object attribute. This is through a new member function called UpdateArsModeAll. It's usage can be seen in the next diff in this stack.

In both cases, the flowset table needs to be checked for availability of space.

In native BCM, this is checked by calling bcmSwitchObjectEcmpDynamicFlowSetFree. Use SAI_SWITCH_ATTR_ARS_AVAILABLE_FLOWS to get this for SAI.

Reviewed By: msomasundaran

Differential Revision: D61363728

fbshipit-source-id: 459dc80934f9a0e5c925d659c651b6078c37f186