fix dhcp test
Summary: DHCP test has destination address which is not rif address. it is a random destination address. it is a unicast DHCP packet. the packet is not a right test case. DHCP packets intended for rif address OR multicast/broadcast DHCP should get trapped. Fixed this test by specifying rif address

Reviewed By: jasmeetbagga

Differential Revision: D67300136

fbshipit-source-id: 895160e51546e0727f188fb034157a8a0e20eb50