Set max switch attribute post create switch on warm boot
Summary:
We learnt of the need to set a non-default max switch id on J3 only
later. And we incorporated it in create switch attributes.
However, BRCM-SAI does not look at create switch attributes on
warm boot. So, to cater to the systems which will only
WB to this change, set the attribute post WB. It should
be a no-op if already set.
Note that this is gated on 12.0 for now, but 11.0 patch is required
and coming.

Reviewed By: shri-khare

Differential Revision:
D67409179

Privacy Context Container: L1125642

fbshipit-source-id: a85f9fff641ae9f470ab677b0a15591913c55114