Restructure config utilities
Summary:
Part of the changes required to have a RuntimeConfig which pulls from TestConfig and PlatformManager config.

Moves cdev_types.py out of /utils/ to prevent circular import loops. Adds a new config.py which distinguishes between the `RuntimeConfig` and `TestConfig` and in the next diff will  hold the logic to generate the RuntimeConfig.

Separates behavior between OSS and FBPKG where FBPKG will have the configs built-in from the configs directory, while OSS runs need to specify a directory to read configs from.

Reviewed By: kimdo8736

Differential Revision: D67216246

fbshipit-source-id: 06eff82c7995b16a35c114d1f1eecff9388249c6