Allow more time for ports to come up after qsfp service cold boot
Summary:
We are currently waiting only 1min for ports to come up after cold-booting qsfp service. This is causing flakiness in tests. This was exposed by D67580376 which optimizes waitForAllCabledPorts implementation.

Our overall expectation is for services to stablize and ports to come up within 5mins. Matching that in the test too

Reviewed By: birdsoup

Differential Revision: D67885054

fbshipit-source-id: 35451c263df0076fe74f2519c7ed988dfdb3b6c6