Add retries for checkNoStats change
Summary:
Main point of this function is to check for continuous
stat changes (for e.g. when a accumulating counter is interpreted
as clear on read), so allow for more than one retry by default.
The main issue seen here is this
- Test checks for stat X
- Stat Y (for e.g. VOQ stats) have still not been updated due to
pkts sent for inducing stat X bump.
- Due to a single retry we end up noticing stat Y change, and end
up failing the test.

TODO: allow for passing in a lambda which asserts for stat X not
changing at all. Since we already asserted for it

Reviewed By: nivinl

Differential Revision:
D67914896

Privacy Context Container: L1125642

fbshipit-source-id: 7e9429b04619a0c514dde88b65ac45e8d16fe57b