Limit number of mac address and neighbor entries to block
Summary: The set mac address to block and neighbors to block thrift apis modify class id of existing entries so that the packets will get dropped due to an acl which matches the class id. Since these apis do not create new entries, we don't need overflow protection as mac and neighbor overflow protection is present. Imposing a limit on the max entries processed. We do not need to account for the entries in swswitch state as the table is created newly for each thrift call.

Reviewed By: shri-khare

Differential Revision: D67953008

fbshipit-source-id: d60a3e960f481511bfdd497b6a17a5cee4a94958