Accumulate linkDownCellDrop counter in agent as its COR in SAI
Summary:
fabricLinkDownDroppedCells_ counter is a clear on read counter from SAI POV. So,
when the counter is read in fboss, unless we do the accumulation, we'll end up with
"monotonic counter fab1/36/8.fabric_link_down_dropped_cells went backwards from 2
to 0" everytime we see a drop in one iteration and lower/no drop in the next. Address
this by doing counter accumulation in fboss.

Reviewed By: jasmeetbagga

Differential Revision: D67918113

fbshipit-source-id: 468251f5fc6dd50a37b2031162e47473234c7ccf