Open source example FBOSS2 CLI command
Summary:
The example code for adding FBOSS2 CLI subcommands is currently private (i.e. located within a `facebook` directory. This change moves the example to be synced to GitHub, and removes some of the facebook internal dependencies (i.e. coop and TBgpService).

Context:
Arista is looking to make a contribution to the FBOSS2 CLI to fetch some data from logs for the purposes of debugging (similar to the "show techsupport" command that is available internally). They've requested that we add a minimal example to OSS as a starting point.

Reviewed By: shri-khare

Differential Revision: D68031014

fbshipit-source-id: 9fd46845b502c3f520e4a1d101454f8ca67ca36f