Open source FBOSS2 CLI documentation for vendor reference
Summary:
Any materials under an "fb" or "facebook" directory are omitted from being synced to GitHub. With D68031014 landing, we are open sourcing a bare minimal exam so that vendors can contribute to the FBOSS2 CLI. This change also open sources the documentation, so that vendors can understand some of the implementation decisions in the CLI, and how add their own logic.

This change also includes a few fixes (not comprehensive) so that the some links will load properly when rendered by GitHub markdown.

Reviewed By: srikrishnagopu

Differential Revision: D68110076

fbshipit-source-id: 2646defd189a4bfff06a531918c8f8c2059275bb