Retry pkg management
Summary:
This handles the case where an RPM is already installed, and we also have initramfs kmods that conflict with the current BSP.

If pkg management fails, retry and force reload. In this situation where this happens we know the RPM must be installed already, as it would have unloaded through the other path if the rpm wasn't installed. This ensures we will have a kmods.json to unload with.

Reviewed By: somasun

Differential Revision: D68279367

fbshipit-source-id: 4bc6cd1df1568c69416660a3849d94b3e6cdad2b