Initial Commit
Summary:
This  is FixMyServer/FixMyBMC for the x86 userver on fboss switches. The main idea is to be able to provide triaging information and quick remediations to common and known  problems that we run into, usually on lab devices. This will reduce turnaround time and pings to fboss_platform team members as most  of the problems that have required manual intervention really could be handled by a tool like this.

Code is identical to that in FixMyBMC except for the actual checks, naming, and some small changes due to how Buck works. In the future I plan on making this more generic so that we can just copy/paste to keep the two libraries in sync.

Reviewed By: tao-ren

Differential Revision: D67463811

fbshipit-source-id: 76a484c85ea95f98e49d70128edde3cc74b6fe58