no transaction support for chenab and always support transaction for tajo
Summary:
rely on asic information available in SwSwitch to decide if transaction can be supported

ChenabAsic has some issue with SAI store reload which is essential part of rollback. This is because certain get APIs are broken. The warm boot support is coming later and rollback may be supported around then. Till then suppressing transactions for Chenab

Reviewed By: jasmeetbagga

Differential Revision: D67808464

fbshipit-source-id: 30f535b5070dde0d4bd9ec6b20cbf99e38e1a850