Parameterize ecmp group and width for benchmark tests
Summary:
As titled. This is in preparation to test 16x2K wide ECMP group supported in 12.x.

Also noticed an issue with the benchmark - we were not modifying ECMP width in the test, hence it is still programming the default ECMP width (64), leading to incorrect benchmark time. Fixing the benchmark.

Reviewed By: maxwindiff

Differential Revision:
D68309761

Privacy Context Container: L1125642

fbshipit-source-id: 077ed35d555b5cf45ef4b9d4eabd390a66367146