Update meru800biab to new xcvr paths
Summary:
Paul reached out to me that this viper is having issue with Led Service.

Turns out it was `meru800biab`. Its XCVR paths weren't updated in D66894266 as it didn't have corresponding csv and cpp file.
But it seems to share with `meru800bia`. So updating its xcvr paths.

Reviewed By: harshitgulati18

Differential Revision: D68460697

fbshipit-source-id: d31d8e2b03e52d1497b1ef7b018f50051b66d20e