Move fmtlib definitions for sai_object_type_t out of gate for SAI >= 1.8.1
Summary:
Issue was reported from CSCO: P1717667627

This specifically occurred when building against SAI 1.7 (did not occur against SAI 1.14). The issue is that these fmtlib definitions are gated (unnecessarily) behind SAI >= 1.8.1 (hence the constexpr expansion by fmtlib for this type fails in the trace when built against 1.7). Moving the definitions out of that gate should resolve it.

Reviewed By: shri-khare, srikrishnagopu

Differential Revision: D68595844

fbshipit-source-id: 29c9c45785a5e58bce2ae25861c8cf63ab810d47