configure port router interface type
Summary:
as titled

in addition, break the loop of finding best profile for a given speed (if max speed is not being picked) as soon as profile is found

this is being done to prevent profile change even if speeds doesn't change. profile change leads to lane change

for chenab port bandwidth test for 400G, initial config was profile 45 (PROFILE_400G_4_PAM4_RS544X2N_COPPER)  but speed setting was profile 10 (PROFILE_400G_8_PAM4_RS544X2N) leading to port re-creation.

port-recreation will not be supported on chenab as it needs to recreation of router interfaces. and port recreation is not required once correct speed is setup.

Reviewed By: msomasundaran

Differential Revision:
D68526445

Privacy Context Container: L1125642

fbshipit-source-id: 6e7591ec101ca541219fedb87133b8ba8436f299