Hold fd for watchdog file
Summary:
FanService manages the fan watchdog. Currently, it opens and closes the sysfs file on every kick/write. Due to the way the watchdog subsystem works, closing a watchdog file attempts to stop the watchdog. However, we use the MAGICCLOSE option which means the watchdog cannot be stopped just by closing the file.
This results in log spew like below:
{F1974766551}

Every time the watchdog is kicked, it tries to stop, fails, and emits a scary-looking log line.

To fix this, have the Bsp class hold a single file descriptor to the watchdog file, and only close it when the service stops.

Reviewed By: somasun

Differential Revision: D68786572

fbshipit-source-id: 3bd09a0444789204f11fe819a4f6c15487266935