Consider drained state of remote port for deciding blinking of LED
Summary:
Port's drained state is used to control blink state of LEDs.
Port is considered drained when the local side is drained (either port or switch) or the peer is drained. The later is handled in this diff

Peer is considered drained if the local side is inactive and has expected reachability.

Reviewed By: xanabani

Differential Revision: D68687303

fbshipit-source-id: 3f496b5e0acffe9d98f56038c3fc298d51a0f637