Eepromv6: Bug fix in getProductionState V6.
Description:-
 The platform manager is reporting an error because it cannot parse the
production state value from the Eepromv6. This is due to an extra space in the
key "Production State " (it should likely be "ProductionState" or
"Production_State").  The parsing logic is failing to recognize the key with
the added space.

Motivation:-
 In V6, the platform manager reports an error because it cannot able to parse
the ProductionState

2578 PlatformExplorer.cpp:211] Exploring Slots for PmUnit NETLAKE at SlotPath /COME_SLOT@0. Count 0
2578 PlatformExplorer.cpp:227] Exploring SlotPath /PDB_SLOT@0
2578 DataStore.cpp:33] Updating bus INCOMING@0 in SlotPath /PDB_SLOT@0 to bus number 23 (i2c-23)
2578 PlatformExplorer.cpp:380] SlotType PmUnitName: PDB
2578 PlatformExplorer.cpp:390] Going with PmUnit name `PDB` defined in config for /PDB_SLOT@0
2578 DataStore.cpp:117] At SlotPath /PDB_SLOT@0, updating to PmUnit PDB with
2578 DataStore.cpp:140] Resolved /PDB_SLOT@0 to default PmUnitConfig of PDB. No ProductSubversion was read from IDPROM at the slotPath.

Testcase:-
The platform manager executes successfully with both V5 and V6 EEPROM formats.
The platform manager executes successfully with both COMe SKU1 and COMe SKU2.
The above test is performed after a power cycle.