Close watchdogs before unloading kmods
Summary:
Following up to D68588373 and https://fb.workplace.com/groups/1419427118405392/permalink/2424591121222315/

In a perfect world the service which kicks the watchdog will always close it. However, in cases where that service crashes and is unable to do cleanup the watchdog will be left open.

If this happens and then you try to re-run platform_manager with "--reload-kmods" you will be completely stuck because unloading the watchdog kmod will unregister the device, leaving it dangling preventing you from fully unloading the kmod ever until a reboot.

To handle this we just close the watchdogs when trying to unload kmods.

This is kind of a hacky solution, and I don't like that the logic lives in PkgManager. I considered moving it to `Utils()` but it also seems bad there as we don't want to just throw everything into a util class.
I imagine the only way to make it significantly less hacky is to do a pretty big restructuring and maybe introduce a BSP interface, but tbh that might be a fool's errand.

Overall it's pretty simple so maybe the concerns are unwarranted.

Reviewed By: birdsoup

Differential Revision: D68931374

fbshipit-source-id: 49e664b2f77f570bf2f0bf51c2a8feac70517f31