Only create transceivers that are present in platform mapping
Summary:
We ran into a qsfp service crash on the dual stage EDSWs where the platform mapping does not enumerate port 19 and 28 because it takes up the permitted VoQ scale.

If the platform mapping doesn't enumerate certain ports, and yet there is an optic detected on that port (shouldn't be but we can't eliminate this situation), then we create the corresponding QsfpModule objects. The contructor of QsfpModule then checks and asserts that there are ports defined in platform mapping for it.

Given that platform mapping hasn't enumerated a port, we shouldn't really have to manage this transceiver. Thus ensuring platform mapping has a port defined before creating QsfpModule objects. Note that this makes the Check() in QsfpModule ctor redundant but keeping it around in case there are bugs at other places that we can address as well.

Reviewed By: xanabani

Differential Revision:
D69078268

Privacy Context Container: L1125642

fbshipit-source-id: 3bfc5729fe672421e297768411bb83e54827e84e