turn on counter_refresh_interval for route scale and pause tests
Summary:
In FBOSS, we have a dedicated  background thread to do stats collection every 1s.
From Leaba 24.8 to 1.42, the time required for a "stats collection" is increased from **500ms to 2.5s** without SDK shadow cache. (more registers are read in each loop, each register access from the hardware involves more hardware access calls.) This is not considered as a prod blocker as we always turn on SDK cache (counter_refresh_interval) in prod, which takes ~ 100 to 200ms for state collection for 24.8.

Route scale test cases (Heavy SAI write) are timeout due to this as SDK shadow cache is disabled by default in testing env.

We are tracking with Cisco to improve this, but the feedback is that this is not easy and might not be improved in a short term.

To fix these failed test cases, current solution is to turn on the SDK cache for these test cases (set counter_refresh_interval).

Reviewed By: vasantspatil

Differential Revision:
D69080434

Privacy Context Container: L1125642

fbshipit-source-id: f0bb2a391faaf0685940bf6cfa73b5534a868447