Introduce New MediaInterfaceCode for FR4_LITE_2x400G
Summary:
**Overview**
This diff introduces a new MediaInterfaceCode for FR4_LITE_2x400G, which is a variant of the base media interface with a distinct characteristic. The key difference between this new code and the base is the SMF length, which is set to 500 meters for FR4_LITE_2x400G (compared to 3k for the base).

**Changes**
Introduced a new MediaInterfaceCode enum value for FR4_LITE_2x400G
Defined the SMF length for FR4_LITE_2x400G as 500 meters
Updated detection logic to identify FR4_LITE_2x400G based on its unique SMF length
All other capabilities remain the same as the base type

Reviewed By: harshitgulati18

Differential Revision: D69364140

fbshipit-source-id: c31dd2cf7cfdc47559cc1e8baf8ff2d9caeacbcd