Create a tcvrID to tcvrName map
Summary:
Later in the diff stack, I plan to add a string tcvrName of the form (eth|fab)X/Y (omitting the /Z) which can be used by downstream clients such as Netstate to create transceiver specific collections. Today those collections are at an interface level, which is causing us to have duplicate collections for all interfaces when the collection only matters at a transceiver level.

In this diff stack, I create a one-time map of tcvrID to tcvrName which will be used later in the diff stack to populate a tcvrName field in TcvrState and TcvrStats

Reviewed By: xanabani

Differential Revision: D69361451

fbshipit-source-id: dd5a550373f9d4d566ae269b5b29082f43a5c11b