scope resolver for interface created on default vlan
Summary: fixing logic to identify scope for vlan rif. look for vlan which has same interface id as that of a given interface id, instead of assuming vlan id is same as interface id. next, if this vlan id is default vlan, then it all l3 switches are in scope for it.

Reviewed By: jasmeetbagga

Differential Revision:
D69285522

Privacy Context Container: L1125642

fbshipit-source-id: bf3f5e96a08372713fb36f86ef4c79c36fdb4606