Introduce minLink per device config
Summary:
Previously, this was needed for J3 only.
Now, we need this for every Virtual Device in R3.

Thus, add new fields with 'PerDevice' in name.
Also, this field is added in SwitchInfo - unique for an NPU.

J3 has single virtual device, so we will set existing threshold as is on J3.

This model can support a mix of J3 and R3 in a single platform as now we can
set these thresholds separately for each NPU.

Reviewed By: jasmeetbagga

Differential Revision:
D69485694

Privacy Context Container: L1125642

fbshipit-source-id: 9bdf5bbcf5da44d392bbbfd02f37e0a809cb0a88