Restructure Serializable remove to use WritableImpl SFINAE template specialization
Summary:
Using WritableImpl SFINAE for template instantiation instead of using method. this borrows the idea from path visitor.

Reason: if we use c++ type, we need to assume the type on the container, for example if it is std::map, F14FastMap etc. Instead, the diff does SFINAE on thrift struct type, which is just map type. But in that case, there is not helper method to reference key or value type. So we need to do some trick to get those types. See diff for detail.

Differential Revision:
D68642459

Privacy Context Container: L1125642

fbshipit-source-id: 3207da89cb32488e8af94ba81c27781823889777