Add message type field in link state notification
Summary:
Changes for S478186: Inconsistent port status on DSF

we see stray linkstate update on port 0 which brings it down.

- There is a port on hw agent which is continously flapping and sdk is providing link state callbacks
- There is continous link active state changes happening as well and sdk is providing active state callbacks
- When hw agent sends the notification to sw agent, the same message is used for link state, link active and link connectivity
```
struct LinkChangeEvent {
  1: optional LinkEvent linkStateEvent;
  2: LinkActiveEvent linkActiveEvents;
  3: LinkConnectivityEvent linkConnectivityEvents;
}

struct LinkEvent {
  1: i32 port;
  2: bool up;
  3: optional phy.LinkFaultStatus iPhyLinkFaultStatus;
  4: switch_config.PortType portType;
}
```

- link state event is an optional field.  Notification from hw agent will have only one of the event types
- When the problem happens, the port field in linkStateEvent is always 0. This leads to suspicion that we are probably reading an unset optional here.

- Thrift IDL documentation lists these kind of optional access as a hazard during deserialization

https://staticdocs.internalfb.com/thrift/docs/idl/field-qualifiers

In C++:
Deserializing does not un-set optional fields which are absent, so deserializing into a single target object multiple times will lead to leftover optional fields from prior deserializations.

The suspicion is that we are reading unset optional field here because the deserializing is happening to a single object and the previous is_set field is getting used. Since the optional linkStateEvent is not present, read for port id results in returning default value of 0.

Fix:
Indicate the message type explicitly in the thrift structure

Reviewed By: daiwei1983

Differential Revision: D69550670

fbshipit-source-id: c9e2233a172a387a1189b3601a16102e1a6802f4