Only program RX settings if they've changed
Summary:
When investigating port flaps, there are a few locations in the code where we might potentially change port settings without any sort of logging. As part of investigating port flaps during this oncall rotation, I'm trying to clean up some of these locations where we lack visibility into potentially disruptive operations.

Add some logic to only change the settings only if they're different, and also log when we actually change the settings so we can see whether this is the cause of a flap

Reviewed By: harshitgulati18

Differential Revision: D69695193

fbshipit-source-id: 9a88c478f2659d080ab59bf8df5ba45426607ff5