Templatizing getInterfaceCode
Summary:
This is the first phase of the changes, where we are done with refactoring CmisModule.cpp without adding AEC specific code yet. In the next few diffs, we will add support for AEC.

For now, we validate nothing is broken, merge this diff stack and then work on Checking-in AEC specific changes in stages.

========

We have a map that translates the speed (e.g. 100G) to a specific Interface code applicable to the transceiver (e.g. SMF CWDM4_100G).

Templatizing the function since the same functionality is required for AEC Interface codes that could have the same number as SMF codes but different encodings (SMF Media Side and AEC Host side interface codes are used, could have the same enum number per but different encodings per Spec: SFF-8024)

Reviewed By: harshitgulati18

Differential Revision: D69302135

fbshipit-source-id: dfb3e6fbeb384ac721355a2584b3dae373bed896