Templatizing getValidMultiportSpeedConfig
Summary:
Given a map of valid speed combinations and module capabilities, we want to figure out if there is a speed combination available that supports the required port speed change.

Templatizing it to avoid duplicating code for AEC.

This will help us in later diffs to avoid duplicating code, and make decisions based on transceiver type (SMF/AEC).

Reviewed By: harshitgulati18

Differential Revision: D69111975

fbshipit-source-id: 430369004c2a7a5947fdf798c6f940c67c908791