Change state update time metric from histogram to quantile
Summary:
Change state update time metric from histogram to quantile. The main advantage of quantile is that it doesn't have a fixed upper bound, so outliers can be recorded accurately.

I noticed there's also a routeUpdate_ stat, but there is no equivalent of TLHistogram::addRepeatedValue() in quantile stats, so leaving it as is for now. Since it's already normalized by number of routes, it's less likely to run into lower/upper bound issues.

Reviewed By: nivinl

Differential Revision: D69962333

fbshipit-source-id: ae3dcb6fc6bf35b03f6df6a595c3c55a7af12086