Change restart time tracker to take service name as an optional parameter
Summary:
Restart time tracker calculates time for each stage of the restart by writing the time to a file. Wedge agent timestamps are written to /dev/shm/fboss/warm_boot.

But for new services which want to calculate restart time instead of creating new directory for each service we will pass service name as a parameter and then save the files in a pre-existing directory like /dev/shm/fboss with the service name and event type concatenated.

Differential Revision: D69958146

fbshipit-source-id: 1d0ca16a34b9c7e796193072a35ed4f32826612e