Update firmware version publishing in PM to use /run/devmap/inforoms
Summary:
Update firmware version publishing in Platform Manager to use /run/devmap/inforoms ; this will fix spurious errors being reported when an FPGA has 2 entries in /run/devmap/fpgas, and only the _INFO_ROM one contains fw_ver, e.g. on Morgan800cc.

As this diff is stacked with the PM and config changes that create the symlinks, it is safe to land without waiting for an intermediate deployment.

Reviewed By: kimdo8736

Differential Revision: D70341321

fbshipit-source-id: d378701172af4da90eff08c5fba772d68a661bf6