GitHub Actions add --local flag (#342)
Summary:
Pull Request resolved: https://github.com/facebook/fboss/pull/342

GitHub actions take care of checking out a specific commit which we want to attempt building (this could be HEAD, or a commit relating to a pull request). We have recently started publishing stable commits that include a pinned revision for FBOSS. In order to build against the desired commit instead of the pinned commit, we need to use the `--local` flag.

Reviewed By: simuthus-fb

Differential Revision: D70211074

fbshipit-source-id: 4894259b8984f7bfe46b316ba18dbd2f39006b89