Add initial documentation for how to use run_test.py
Summary: As requested by vendors - most of the information on how to use `run_test.py` currently lives in the script itself. This diff starts some documentation with examples on how to use the script and some special flags for QSFP hw tests and link tests. It also includes the recent changes we made to support new flags for QSFP and link tests (see D70271214 and D70348574).

Reviewed By: harshitgulati18

Differential Revision: D70508027

fbshipit-source-id: b1e8c0f250dadf50f373e001e48297d7f50940c6