Avoid access into SDK in PFC deadlock notification callback
Summary:
Currently, we are getting the Index and Port attributes of queue based on the QueueSaiId
passed in the deadlock detection callback from SDK. This call is entering SDK which is
causing a deadlock in KODIAK3 implementation. The recommendation is to avoid calling
back into SDK in a callback from SDK, which is exactly whats happening here. To avoid
this, we can either go with top half/bottom half approach or avoid a call into SDK. Given
the processing we are doing for PFC WD detection callback is pretty static, we can avoid
the overhead of this split and fix the issue by getting the needed information from SAI
store instead.

Reviewed By: vasantspatil, maxwindiff

Differential Revision: D70467119

fbshipit-source-id: 15e7cf06778476a7420080af4a909ff0c72082e5