Method to get DSCP mode
Summary:
As titled, Cisco doesn't support PIPE mode for DSCP. We have requested this
feature. Unitl then we are going to enable this feature in prod using UNIFORM
mode. Confirmed with the team that this shouldn't matter as we update inner and
outer DSCPs ourselves.

Differential Revision:
D69792884

Privacy Context Container: L1125642

fbshipit-source-id: d58680ab848a0953a3e4442d7f66970f65570f92