Don't reset SHEL attributes if no SHEL config
Summary:
This is towards supporting SHEL config change/removal.

When SHEL is disable, we would try to reset the SHEL attributes to default value (e.g. ShelSrcIp to default route). This does not make sense and also rejected by the SDK.

Since this is static configuration, we can leave it in the switch and disable port shel enable and system port shel destination enable. That would essentially disable SHEL functionality.

Reviewed By: jasmeetbagga

Differential Revision: D70362284

fbshipit-source-id: ced5022f6993b8382203a28343f8531b285f68e6