Refactor PortTxEnableTest to be more reliable
Summary:
Previous version seemed to be racy, it was checking
for packets stoppage by querying comparing 2 back
to back stats collection and asserting that packets
did not increment. But it could well be the case that
packet stat simply wasn't incremented yet.

Was seeing this consistently fail on J3 warm boot iterations.

Instead simplify the test to do the following
- Setup data plane loop on 2 ports
- Assert for atleast 1Gbps traffic on both
- Enable Tx pause on first port and assert that traffic drops to 0. When
asserting this, assert that in the same stat collection the other port
continues to see increment. This negates the race seen above.
- Now enable tx on first port
- Assert that high traffic rate can again be achieved on both ports

Reviewed By: nivinl, maxwindiff

Differential Revision: D70600799

fbshipit-source-id: 03f8e4cd8470a43f1730bebe4e75f6333f9db17b