Do not rely on file presence for cpp wrapper determination
Summary:
In S494101, sometimes agent is failing to start during re provisioning because the fboss_sw_agent service symlink is not setup. The symlink is set up by agent pre executer. The logs show that pre-executer ran and the systemd moved to the agent executer which tries to start the sw agent.

Based on code analysis, this can happen if the pre-executer incorrectly determines that it is running python wrapper and not the cpp wrapper. This determination is done by checking the presence of the feature file /etc/fboss/features/cpp_wedge_agent_wrapper/current/on. The suspicion is that the utility that checks for presence of this file is failing in the initial attempt even though the file is present. A systemctl start of the process subsequently fixes the symlink.

List of changes:
- use agent config to determine whether cpp wrapper is enabled
- add debugs which will help to identify this situation

Reviewed By: phshaikh

Differential Revision: D70650938

fbshipit-source-id: f94a4878cc97af24dbde8062fe3081bf0dbcab11