Expect UNIFORM mode DSCP value in packet
Summary:
As titled.
In morgan800cc, supported tunnel DSCP mode is UNIFORM. Where as in other platforms it is PIPE. Please see the description about these two mode in SAI spec: https://github.com/opencomputeproject/SAI/blob/master/inc/saitunnel.h

Accordingly, we changed the expectation in test case for Cisco.

Note that in the test case outer traffic class value is 0xFA ==b'1111 1010'. But, expected traffic value (after trapping inner packet) is 0xF8 == b'1111 1000' Because ECN bits are not copied.

But when we created the packet we added inner traffic class field value as 0xCE == b'1100 1110' So, even if we replace the DSCP bits, ECN bits should still be preserved and we should have received, 0xFA? wondering why it's not preserved. Will check this with Cisco.

Differential Revision: D70483871

fbshipit-source-id: 3a559952867ca0794132a2c909d896862d759a9d