Update fw_ver paths in fw_util configs to /run/devmap/inforoms
Summary:
Update fw_util configs to use new symlinks for inforoms; since fw_util config already hardcodes these paths, there are no ambiguity issues like PM had, but it's better to keep the fw_ver paths used by PM and fw_util in sync. In the future we can consolidate the logic for reading firmware versions into a common library for fw_util and PM.

fw_util is always deployed as part of platform_stack, so this is safe with the previous PM diff in this stack.

Reviewed By: joancaneus

Differential Revision: D70341323

fbshipit-source-id: ca5e478ecd7d3d776a95238b8ef304786be0b9d7