Remove github_actions_fboss_platform_services target
Summary: as titled. We're removing the `github_actions_fboss_platform_services` target as it overlaps with `fboss_platform_services` after D70916665. We originally had two different targets as I suspected adding additional binaries to be built during on-diff would be detrimental to on-diff build time, but build time with these new binaries added seems to still be around the 30min mark, and I think there is value in the added coverage.

Reviewed By: somasun

Differential Revision: D70917563

fbshipit-source-id: 929949c9b91e692b2ab20a424cda43f914edad16